Get rid of BrowseItemPanel #4613
-Removed BrowseItemsSelectionPanel (and inheritors ) as this panel is not to be shown anymore;
