+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/content-studio/js/app/browse/ContentBrowseItemPanel.ts
@@ -1,13 +1,7 @@
 import '../../api.ts';
-import {ContentBrowseItemsSelectionPanel} from './ContentBrowseItemsSelectionPanel';
 import {ContentItemStatisticsPanel} from '../view/ContentItemStatisticsPanel';
-import {ContentTreeGrid} from './ContentTreeGrid';
 
 export class ContentBrowseItemPanel extends api.app.browse.BrowseItemPanel<api.content.ContentSummaryAndCompareStatus> {
-
-    createItemSelectionPanel(grid: ContentTreeGrid): ContentBrowseItemsSelectionPanel {
-        return new ContentBrowseItemsSelectionPanel(grid);
-    }
 
     createItemStatisticsPanel(): ContentItemStatisticsPanel {
         return new ContentItemStatisticsPanel();
