+++ modules/admin/admin-ui/src/main/resources/web/admin/common/js/app/browse/_module.ts
@@ -1,7 +1,6 @@
 ///<reference path='BrowsePanel.ts' />
 ///<reference path='BrowseItem.ts' />
 ///<reference path='SelectionItem.ts' />
-///<reference path='BrowseItemsSelectionPanel.ts' />
 ///<reference path='BrowseItemPanel.ts' />
 ///<reference path='ItemDeselectedEvent.ts' />
 ///<reference path='BrowseItemsChanges.ts' />
