--- modules/admin/admin-ui/src/main/resources/web/admin/common/styles/api/app/browse/items-selection-panel.less
@@ -1,69 +1,0 @@
-.items-selection-panel {
-
-  &.no-selection {
-    .watermark-text(21px, true);
-    margin-top: -15px;
-
-    .toolbar {
-      display: none;
-    }
-  }
-
-  .items-container {
-    position: static;
-    top: 0;
-    bottom: 0;
-    width: 100%;
-    box-sizing: border-box;
-    max-height: 100%;
-    padding-bottom: 15px;
-    overflow-x: hidden;
-    overflow-y: auto;
-
-    .browse-selection-item {
-      min-width: 100px;
-      max-width: 200px;
-    }
-  }
-
-  &:not(.no-selection) {
-    .items-container {
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: flex-start;
-    }
-  }
-
-  &.no-toolbar {
-    .show-all, .fold-button {
-      display: none !important;
-    }
-  }
-
-  .toolbar {
-    background-color: @admin-bg-almost-white;
-    border-top: 1px solid @admin-bg-light-gray;
-    text-align: center;
-
-    .@{_COMMON_PREFIX}button {
-      &:active, span {
-        color: @admin-button-blue2;
-      }
-
-      &:hover:not(:disabled) {
-        span {
-          color: @admin-button-blue3;
-        }
-      }
-
-      &:focus {
-        box-shadow: none;
-        border: none;
-      }
-    }
-
-    .@{_COMMON_PREFIX}dropdown {
-      bottom: 32px;
-    }
-  }
-}
