+++ modules/admin/admin-ui/src/main/resources/web/admin/common/styles/api/ui/panel/split-panel.less
@@ -10,10 +10,6 @@
     cursor: ns-resize;
   }
 
-  .panel.items-selection-panel {
-    overflow-y: auto;
-  }
-
   > .splitter {
     background-color: @admin-bg-light-gray;
     cursor: ns-resize;
