--- modules/admin/admin-ui/src/main/resources/web/admin/common/styles/apps/content/browse/content-browse-items-selection-panel.less
@@ -1,10 +1,0 @@
-.content-browse-items-selection-panel {
-
-  margin-bottom: 42px;
-  overflow: visible !important;
-
-  &.no-selection {
-    margin-bottom: 0px;
-  }
-
-}
