+++ modules/admin/admin-ui/src/main/resources/web/admin/common/styles/styles.less
@@ -79,7 +79,6 @@
 @import "api/app/appbar-tab-menu";
 @import "api/app/browse/browse-item-panel";
 @import "api/app/browse/browse-selection-item";
-@import "api/app/browse/items-selection-panel";
 @import "api/app/browse/item-statistics-panel";
 @import "api/app/browse/browse-filter-panel";
 @import "api/app/browse/grid/tree-grid-panel";
@@ -134,7 +133,6 @@
 // apps: content
 @import "apps/content/browse/content-browse-toolbar";
 @import "apps/content/browse/content-publish-menu-button";
-@import "apps/content/browse/content-browse-items-selection-panel";
 @import "apps/content/view/content-item-preview-panel";
 @import "apps/content/view/content-item-statistics-panel";
 @import "apps/content/view/mobile-content-item-statistics-panel";
