+++ modules/app/app-applications/src/main/js/app/browse/ApplicationBrowsePanel.ts
@@ -42,7 +42,7 @@ export class ApplicationBrowsePanel extends api.app.browse.BrowsePanel<Applicati
     }
 
     protected createBrowseItemPanel(): ApplicationBrowseItemPanel {
-        return new ApplicationBrowseItemPanel(this.treeGrid);
+        return new ApplicationBrowseItemPanel();
     }
 
     treeNodesToBrowseItems(nodes: TreeNode<Application>[]): BrowseItem<Application>[] {
